Move Sample to common
