+++ common/solar_common/__init__.py
@@ -1,1 +1,2 @@
 from .solar_common import fieldnames
+from .solar_common import Sample
