+++ common/solar_common/solar_common.py
@@ -15,3 +15,26 @@ fieldnames = [
     "load power",
     "battery percentage",
 ]
+
+
+class Sample(
+    TypedDict(
+        "Sample",
+        {
+            timestamp: float,
+            "PV voltage": float,
+            "PV current": float,
+            "PV power L": float,
+            "PV power H": float,
+            "battery voltage": float,
+            "battery current": float,
+            "battery power L": float,
+            "battery power H": float,
+            "load voltage": float,
+            "load current": float,
+            "load power": float,
+            "battery percentage": float,
+        },
+    )
+):
+    pass
